HoopTrack/scripts/autolabel/requirements.txt
torch>=2.1
torchvision>=0.16
transformers>=4.37
pillow>=10.0
tqdm>=4.66
jinja2>=3.1
